@@ -1,16 +1,26 @@
 plugins {
     id 'com.github.johnrengelman.shadow' version '8.1.1' apply false
     id 'maven-publish'
+    id 'java'
+}
+
+jar {
+    manifest {
+        attributes("paperweight-mappings-namespace": "spigot")
+    }
 }
 
 allprojects {
     apply plugin: 'java-library'
     apply plugin: 'com.github.johnrengelman.shadow'
     group = 'dev.rosewood'
-    version = '1.5.20'
+    version = '1.5.21-SNAPSHOT'
+    sourceCompatibility = JavaVersion.VERSION_17
+    targetCompatibility = JavaVersion.VERSION_17
 
     java {
-        toolchain.languageVersion.set(JavaLanguageVersion.of(17))
+        toolchain.languageVersion.set(JavaLanguageVersion.of(21))
+        disableAutoTargetJvm()
     }
 
     repositories {
@@ -19,6 +29,7 @@ allprojects {
         maven { url = 'https://repo.rosewooddev.io/repository/public/' }
         maven { url = 'https://libraries.minecraft.net' }
         maven { url = 'https://repo.papermc.io/repository/maven-public/' }
+        maven { url = 'https://s01.oss.sonatype.org/content/repositories/snapshots/' }
         maven { url = 'https://hub.spigotmc.org/nexus/content/repositories/snapshots/' }
         maven { url = 'https://repo.bg-software.com/repository/api/' }
         maven { url = 'https://bitbucket.org/kangarko/libraries/raw/master' }
@@ -62,6 +73,7 @@ dependencies {
     implementation project(path: ':NMS:v1_20_R2', configuration: 'remapped')
     */
     implementation project(path: ':NMS:v1_20_R3', configuration: 'remapped')
+    implementation project(path: ':NMS:v1_20_R4', configuration: 'remapped')
 }
 
 shadowJar {
