@@ -1,11 +1,11 @@
 package dev.rosewood.rosestacker.command.argument;
 
 import dev.rosewood.rosegarden.RosePlugin;
-import dev.rosewood.rosegarden.command.framework.ArgumentParser;
-import dev.rosewood.rosegarden.command.framework.RoseCommandArgumentHandler;
-import dev.rosewood.rosegarden.command.framework.RoseCommandArgumentInfo;
+import dev.rosewood.rosegarden.command.framework.Argument;
+import dev.rosewood.rosegarden.command.framework.ArgumentHandler;
+import dev.rosewood.rosegarden.command.framework.CommandContext;
+import dev.rosewood.rosegarden.command.framework.InputIterator;
 import dev.rosewood.rosegarden.utils.StringPlaceholders;
-import dev.rosewood.rosestacker.command.argument.StackedBlockAmountArgumentHandler.StackedBlockAmount;
 import dev.rosewood.rosestacker.manager.StackSettingManager;
 import dev.rosewood.rosestacker.stack.settings.BlockStackSettings;
 import java.util.Arrays;
@@ -13,28 +13,29 @@ import java.util.Collections;
 import java.util.List;
 import org.bukkit.Material;
 
-public class StackedBlockAmountArgumentHandler extends RoseCommandArgumentHandler<StackedBlockAmount> {
+public class StackedBlockAmountArgumentHandler extends ArgumentHandler<Integer> {
+
+    private final RosePlugin rosePlugin;
 
     public StackedBlockAmountArgumentHandler(RosePlugin rosePlugin) {
-        super(rosePlugin, StackedBlockAmount.class);
+        super(Integer.class);
+
+        this.rosePlugin = rosePlugin;
     }
 
     @Override
-    protected StackedBlockAmount handleInternal(RoseCommandArgumentInfo argumentInfo, ArgumentParser argumentParser) throws HandledArgumentException {
-        String input = argumentParser.next();
+    public Integer handle(CommandContext context, Argument argument, InputIterator inputIterator) throws HandledArgumentException {
+        String input = inputIterator.next();
         try {
-            return new StackedBlockAmount(Integer.parseInt(input));
+            return Integer.parseInt(input);
         } catch (Exception e) {
             throw new HandledArgumentException("argument-handler-stackamount", StringPlaceholders.of("input", input));
         }
     }
 
     @Override
-    protected List<String> suggestInternal(RoseCommandArgumentInfo argumentInfo, ArgumentParser argumentParser) {
-        String previous = argumentParser.previous();
-        argumentParser.next();
-
-        Material blockType = previous == null ? null : Material.matchMaterial(previous);
+    public List<String> suggest(CommandContext context, Argument argument, String[] args) {
+        Material blockType = context.get(Material.class);
         if (blockType == null)
             return Collections.singletonList("<stackSize>");
 
@@ -45,7 +46,5 @@ public class StackedBlockAmountArgumentHandler extends RoseCommandArgumentHandle
         int maxStackAmount = blockStackSettings.getMaxStackSize();
         return Arrays.asList(String.valueOf(maxStackAmount), String.valueOf(maxStackAmount / 2), String.valueOf(maxStackAmount / 4), "<stackSize>");
     }
-
-    public record StackedBlockAmount(int amount) { }
 
 }
