@@ -1,11 +1,11 @@
 package dev.rosewood.rosestacker.command.argument;
 
 import dev.rosewood.rosegarden.RosePlugin;
-import dev.rosewood.rosegarden.command.framework.ArgumentParser;
-import dev.rosewood.rosegarden.command.framework.RoseCommandArgumentHandler;
-import dev.rosewood.rosegarden.command.framework.RoseCommandArgumentInfo;
+import dev.rosewood.rosegarden.command.framework.Argument;
+import dev.rosewood.rosegarden.command.framework.ArgumentHandler;
+import dev.rosewood.rosegarden.command.framework.CommandContext;
+import dev.rosewood.rosegarden.command.framework.InputIterator;
 import dev.rosewood.rosegarden.utils.StringPlaceholders;
-import dev.rosewood.rosestacker.command.argument.StackedEntityAmountArgumentHandler.StackedEntityAmount;
 import dev.rosewood.rosestacker.manager.StackSettingManager;
 import dev.rosewood.rosestacker.stack.settings.EntityStackSettings;
 import java.util.Arrays;
@@ -13,28 +13,29 @@ import java.util.Collections;
 import java.util.List;
 import org.bukkit.entity.EntityType;
 
-public class StackedEntityAmountArgumentHandler extends RoseCommandArgumentHandler<StackedEntityAmount> {
+public class StackedEntityAmountArgumentHandler extends ArgumentHandler<Integer> {
+
+    private final RosePlugin rosePlugin;
 
     public StackedEntityAmountArgumentHandler(RosePlugin rosePlugin) {
-        super(rosePlugin, StackedEntityAmount.class);
+        super(Integer.class);
+
+        this.rosePlugin = rosePlugin;
     }
 
     @Override
-    protected StackedEntityAmount handleInternal(RoseCommandArgumentInfo argumentInfo, ArgumentParser argumentParser) throws HandledArgumentException {
-        String input = argumentParser.next();
+    public Integer handle(CommandContext context, Argument argument, InputIterator inputIterator) throws HandledArgumentException {
+        String input = inputIterator.next();
         try {
-            return new StackedEntityAmount(Integer.parseInt(input));
+            return Integer.parseInt(input);
         } catch (Exception e) {
             throw new HandledArgumentException("argument-handler-stackamount", StringPlaceholders.of("input", input));
         }
     }
 
     @Override
-    protected List<String> suggestInternal(RoseCommandArgumentInfo argumentInfo, ArgumentParser argumentParser) {
-        String previous = argumentParser.previous();
-        argumentParser.next();
-
-        EntityType entityType = previous == null ? null : Arrays.stream(EntityType.values()).filter(x -> x.name().equalsIgnoreCase(previous)).findFirst().orElse(null);
+    public List<String> suggest(CommandContext context, Argument argument, String[] args) {
+        EntityType entityType = context.get(EntityType.class);
         if (entityType == null)
             return Collections.singletonList("<stackSize>");
 
@@ -45,7 +46,5 @@ public class StackedEntityAmountArgumentHandler extends RoseCommandArgumentHandl
         int maxStackAmount = spawnerStackSettings.getMaxStackSize();
         return Arrays.asList(String.valueOf(maxStackAmount), String.valueOf(maxStackAmount / 2), String.valueOf(maxStackAmount / 4), "<stackSize>");
     }
-
-    public record StackedEntityAmount(int amount) { }
 
 }
