@@ -1,36 +1,36 @@
 package dev.rosewood.rosestacker.command.argument;
 
 import dev.rosewood.rosegarden.RosePlugin;
-import dev.rosewood.rosegarden.command.framework.ArgumentParser;
-import dev.rosewood.rosegarden.command.framework.RoseCommandArgumentHandler;
-import dev.rosewood.rosegarden.command.framework.RoseCommandArgumentInfo;
+import dev.rosewood.rosegarden.command.framework.Argument;
+import dev.rosewood.rosegarden.command.framework.ArgumentHandler;
+import dev.rosewood.rosegarden.command.framework.CommandContext;
+import dev.rosewood.rosegarden.command.framework.InputIterator;
 import dev.rosewood.rosegarden.utils.StringPlaceholders;
-import dev.rosewood.rosestacker.command.argument.TranslationLocaleArgumentHandler.TranslationLocale;
 import dev.rosewood.rosestacker.manager.LocaleManager;
 import java.util.List;
 
-public class TranslationLocaleArgumentHandler extends RoseCommandArgumentHandler<TranslationLocale> {
+public class TranslationLocaleArgumentHandler extends ArgumentHandler<String> {
+
+    private final RosePlugin rosePlugin;
 
     public TranslationLocaleArgumentHandler(RosePlugin rosePlugin) {
-        super(rosePlugin, TranslationLocale.class);
+        super(String.class);
+
+        this.rosePlugin = rosePlugin;
     }
 
     @Override
-    protected TranslationLocale handleInternal(RoseCommandArgumentInfo argumentInfo, ArgumentParser argumentParser) throws HandledArgumentException {
-        String input = argumentParser.next();
+    public String handle(CommandContext context, Argument argument, InputIterator inputIterator) throws HandledArgumentException {
+        String input = inputIterator.next();
         return this.rosePlugin.getManager(LocaleManager.class).getPossibleTranslationLocales().stream()
                 .filter(x -> x.equalsIgnoreCase(input))
-                .map(TranslationLocale::new)
                 .findFirst()
                 .orElseThrow(() -> new HandledArgumentException("argument-handler-translationlocale", StringPlaceholders.of("input", input)));
     }
 
     @Override
-    protected List<String> suggestInternal(RoseCommandArgumentInfo argumentInfo, ArgumentParser argumentParser) {
-        argumentParser.next();
+    public List<String> suggest(CommandContext context, Argument argument, String[] args) {
         return this.rosePlugin.getManager(LocaleManager.class).getPossibleTranslationLocales();
     }
-
-    public record TranslationLocale(String name) { }
 
 }
