@@ -1,19 +1,21 @@
 package dev.rosewood.rosestacker.command.command;
 
 import dev.rosewood.rosegarden.RosePlugin;
+import dev.rosewood.rosegarden.command.argument.ArgumentHandlers;
+import dev.rosewood.rosegarden.command.framework.ArgumentsDefinition;
+import dev.rosewood.rosegarden.command.framework.BaseRoseCommand;
 import dev.rosewood.rosegarden.command.framework.CommandContext;
-import dev.rosewood.rosegarden.command.framework.RoseCommand;
-import dev.rosewood.rosegarden.command.framework.RoseCommandWrapper;
+import dev.rosewood.rosegarden.command.framework.CommandInfo;
 import dev.rosewood.rosegarden.command.framework.annotation.RoseExecutable;
 import dev.rosewood.rosegarden.utils.StringPlaceholders;
 import dev.rosewood.rosestacker.manager.LocaleManager;
 import dev.rosewood.rosestacker.manager.StackManager;
 import dev.rosewood.rosestacker.utils.StackerUtils;
 
-public class ClearallCommand extends RoseCommand {
+public class ClearallCommand extends BaseRoseCommand {
 
-    public ClearallCommand(RosePlugin rosePlugin, RoseCommandWrapper parent) {
-        super(rosePlugin, parent);
+    public ClearallCommand(RosePlugin rosePlugin) {
+        super(rosePlugin);
     }
 
     @RoseExecutable
@@ -40,18 +42,18 @@ public class ClearallCommand extends RoseCommand {
     }
 
     @Override
-    protected String getDefaultName() {
-        return "clearall";
+    protected CommandInfo createCommandInfo() {
+        return CommandInfo.builder("clearall")
+                .descriptionKey("command-clearall-description")
+                .permission("rosestacker.clearall")
+                .build();
     }
 
     @Override
-    public String getDescriptionKey() {
-        return "command-clearall-description";
-    }
-
-    @Override
-    public String getRequiredPermission() {
-        return "rosestacker.clearall";
+    protected ArgumentsDefinition createArgumentsDefinition() {
+        return ArgumentsDefinition.builder()
+                .required("type", ArgumentHandlers.forEnum(ClearallType.class))
+                .build();
     }
 
     public enum ClearallType {
