@@ -19,7 +19,9 @@ import dev.rosewood.rosestacker.utils.ItemUtils;
 import dev.rosewood.rosestacker.utils.StackerUtils;
 import dev.rosewood.rosestacker.utils.ThreadUtils;
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.List;
+import java.util.Map;
 import org.bukkit.Bukkit;
 import org.bukkit.GameMode;
 import org.bukkit.Location;
@@ -309,15 +311,23 @@ public class BlockListener implements Listener {
         if (amount <= 0)
             return true;
 
-        if (Setting.SPAWNER_DROP_TO_INVENTORY.getBoolean())
-            dropLocation = player.getLocation().add(0, 1, 0);
-
+        boolean dropToInventory = Setting.SPAWNER_DROP_TO_INVENTORY.getBoolean();
         StackManager stackManager = this.rosePlugin.getManager(StackManager.class);
+        List<ItemStack> itemsToDrop;
         if (Setting.SPAWNER_BREAK_ENTIRE_STACK_INTO_SEPARATE.getBoolean()) {
-            ItemStack spawnerItem = ItemUtils.getSpawnerAsStackedItemStack(spawnerType, 1);
-            stackManager.dropItemStack(spawnerItem, amount, dropLocation, true);
+            ItemStack item = ItemUtils.getSpawnerAsStackedItemStack(spawnerType, 1);
+            item.setAmount(amount);
+            itemsToDrop = List.of(item);
         } else {
-            stackManager.preStackItems(List.of(ItemUtils.getSpawnerAsStackedItemStack(spawnerType, amount)), dropLocation);
+            itemsToDrop = List.of(ItemUtils.getSpawnerAsStackedItemStack(spawnerType, amount));
+        }
+
+        if (dropToInventory) {
+            Collection<ItemStack> remainingItems = player.getInventory().addItem(itemsToDrop.toArray(ItemStack[]::new)).values();
+            if (!remainingItems.isEmpty())
+                stackManager.preStackItems(remainingItems, player.getLocation().add(0, 1, 0));
+        } else {
+            stackManager.preStackItems(itemsToDrop, dropLocation);
         }
 
         return true;
