@@ -15,6 +15,7 @@ import dev.rosewood.rosestacker.stack.StackedEntity;
 import dev.rosewood.rosestacker.stack.settings.conditions.entity.StackConditions;
 import dev.rosewood.rosestacker.utils.PersistentDataUtils;
 import dev.rosewood.rosestacker.utils.StackerUtils;
+import dev.rosewood.rosestacker.utils.VersionUtils;
 import java.lang.reflect.Type;
 import java.util.ArrayList;
 import java.util.EnumSet;
@@ -89,6 +90,7 @@ public class EntityStackSettings extends StackSettings {
             this.stackConditions.add(new StackConditionEntry<>(stackCondition));
 
         this.extraSettings = new HashMap<>();
+
         switch (this.entityType) {
             case CHICKEN -> this.putSetting(CHICKEN_MULTIPLY_EGG_DROPS_BY_STACK_SIZE, true);
             case SHEEP -> {
@@ -96,10 +98,11 @@ public class EntityStackSettings extends StackSettings {
                 this.putSetting(SHEEP_PERCENTAGE_OF_WOOL_TO_REGROW_PER_GRASS_EATEN, 25.0);
             }
             case SLIME, MAGMA_CUBE -> this.putSetting(SLIME_ACCURATE_DROPS_WITH_KILL_ENTIRE_STACK_ON_DEATH, true);
-            case MUSHROOM_COW -> {
-                this.putSetting(MOOSHROOM_DROP_ADDITIONAL_MUSHROOMS_FOR_EACH_COW_IN_STACK, true);
-                this.putSetting(MOOSHROOM_EXTRA_MUSHROOMS_PER_COW_IN_STACK, 5);
-            }
+        }
+
+        if (this.entityType == VersionUtils.MOOSHROOM) {
+            this.putSetting(MOOSHROOM_DROP_ADDITIONAL_MUSHROOMS_FOR_EACH_COW_IN_STACK, true);
+            this.putSetting(MOOSHROOM_EXTRA_MUSHROOMS_PER_COW_IN_STACK, 5);
         }
 
         this.setDefaults();
