@@ -35,8 +35,8 @@ import org.bukkit.permissions.PermissionAttachmentInfo;
 public final class StackerUtils {
 
     public static final String MIN_SUPPORTED_VERSION = "1.16.5";
-    public static final String MAX_SUPPORTED_VERSION = "1.20.4";
-    public static final String MAX_SUPPORTED_LOCALE_VERSION = "1.20.4";
+    public static final String MAX_SUPPORTED_VERSION = "1.20.5";
+    public static final String MAX_SUPPORTED_LOCALE_VERSION = "1.20.5";
 
     public static final int ASSUMED_ENTITY_VISIBILITY_RANGE = 75 * 75;
 
